+++ ehealth-recipe-api/src/main/java/com/ngari/recipe/recipereportform/model/RecipeMonthAccountCheckResponse.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 public class RecipeMonthAccountCheckResponse implements Serializable{
     private static final long serialVersionUID = 8599154333590559939L;
 
-    private Integer total; //总计
+    private Long total; //总计
 
     private Integer organId; //机构id
